Move admin level toggle into mapSettingsMenu folder
